openhost_system_agent/pyproject.toml
[project]
name = "openhost-system-agent"
version = "0.1.0"
description = "OpenHost system agent — host-level updates and migrations"
requires-python = ">=3.12"
dependencies = [
    "attrs",
    "cappa",
    "GitPython",
    "loguru",
]

[project.scripts]
openhost_system_agent = "openhost_system_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openhost_system_agent"]

[tool.uv]
managed = false
